wip: content arrange with prompts
