+++ src/dotnet/MLSearch.Service/Indexing/ChatContent/ChatContentArranger2.cs
@@ -1,4 +1,3 @@
-using System.Text;
 using ActualChat.Chat;
 using ActualChat.MLSearch.Documents;
 
@@ -6,8 +5,8 @@ namespace ActualChat.MLSearch.Indexing.ChatContent;
 
 internal sealed class ChatContentArranger2(
     IChatsBackend chatsBackend,
-    IAuthorsBackend authorsBackend,
-    FragmentContinuationSelector selector
+    IDialogFragmentAnalyzer fragmentAnalyzer,
+    ChatDialogFormatter chatDialogFormatter
 ) : IChatContentArranger
 {
     public async IAsyncEnumerable<SourceEntries> ArrangeAsync(
@@ -30,7 +29,7 @@ internal sealed class ChatContentArranger2(
                     .ConfigureAwait(false);
                 builder.Entries.AddRange(tailEntries);
                 builder.HasInitializedEntries = true;
-                builder.Dialog = await BuildUpDialog(tailEntries).ConfigureAwait(false);
+                builder.Dialog = await chatDialogFormatter.BuildUpDialog(tailEntries).ConfigureAwait(false);
             }
         }
 
@@ -47,7 +46,7 @@ internal sealed class ChatContentArranger2(
                 };
                 builders.Add(builder);
                 builder.Entries.Add(entry);
-                builder.Dialog = await BuildUpDialog(builder.Entries).ConfigureAwait(false);
+                builder.Dialog = await chatDialogFormatter.BuildUpDialog(builder.Entries).ConfigureAwait(false);
             }
             else {
                 // Do we need to close any builder: have some content and last entry created earlier than 1 day from the current one.
@@ -56,10 +55,10 @@ internal sealed class ChatContentArranger2(
 
                 foreach (var builder in builders) {
                     var dialog = builder.Dialog;
-                    var entryText = await EntryToText(entry, builder.Entries[^1]).ConfigureAwait(false);
+                    var entryText = await chatDialogFormatter.EntryToText(entry, builder.Entries[^1]).ConfigureAwait(false);
                     dialog = dialog + Environment.NewLine + entryText;
                     builder.PossibleDialog = dialog;
-                    var result = await selector.IsFragmentAboutTheSameTopic(dialog).ConfigureAwait(false);
+                    var result = await fragmentAnalyzer.IsDialogAboutTheSameTopic(dialog).ConfigureAwait(false);
                     if (result is { HasValue: true, Value: true })
                         candidates.Add(builder);
                 }
@@ -67,7 +66,7 @@ internal sealed class ChatContentArranger2(
                     builderToAdd = candidates[0];
                 else if (candidates.Count > 1) {
                     var dialogs = candidates.Select(c => c.PossibleDialog).ToArray();
-                    var index = await selector.ChooseMoreProbableDialog(dialogs).ConfigureAwait(false);
+                    var index = await fragmentAnalyzer.ChooseMoreProbableDialog(dialogs).ConfigureAwait(false);
                     if (index >= 0)
                         builderToAdd = candidates[index];
                 }
@@ -83,7 +82,7 @@ internal sealed class ChatContentArranger2(
                         HasModified = true
                     };
                     builder.Entries.Add(entry);
-                    builder.Dialog = await BuildUpDialog(builder.Entries).ConfigureAwait(false);
+                    builder.Dialog = await chatDialogFormatter.BuildUpDialog(builder.Entries).ConfigureAwait(false);
                     builders.Add(builder);
                 }
             }
@@ -93,64 +92,6 @@ internal sealed class ChatContentArranger2(
             if (builder.HasModified)
                 yield return new SourceEntries(null, null, builder.Entries);
         }
-    }
-
-    private async Task<string> BuildUpDialog(IReadOnlyList<ChatEntry> chatEntries)
-    {
-        var sb = new StringBuilder();
-        ChatEntry? prevChatEntry = null;
-        foreach (var chatEntry in chatEntries) {
-            if (sb.Length > 0)
-                sb.AppendLine();
-            var entryText = await EntryToText(chatEntry, prevChatEntry).ConfigureAwait(false);
-            sb.Append(entryText);
-        }
-        return sb.ToString();
-    }
-
-    private async Task<string> EntryToText(ChatEntry entry, ChatEntry? prevChatEntry)
-    {
-        var isBlockStart = IsBlockStart(prevChatEntry, entry);
-        var isReply = entry.RepliedEntryLocalId is not null;
-        var text = await ContentToText(entry.Content).ConfigureAwait(false);
-        var showAuthor = isBlockStart || isReply;
-        if (!showAuthor)
-            return text;
-
-        var authorName = await GetAuthorName(entry.AuthorId).ConfigureAwait(false);
-        var timestamp = entry.BeginsAt.ToDateTime();
-        var sTimestamp = $"{timestamp.ToShortDateString()} at {timestamp.ToShortTimeString()}";
-
-        var sb = new StringBuilder();
-        sb.AppendLine(authorName);
-        sb.AppendLine(sTimestamp);
-        sb.Append(text);
-        return sb.ToString();
-    }
-
-    private async Task<string> GetAuthorName(AuthorId authorId)
-    {
-        var author = await authorsBackend
-            .Get(authorId.ChatId, authorId, AuthorsBackend_GetAuthorOption.Full, default)
-            .ConfigureAwait(false);
-        var authorName = author?.Avatar.Name ?? "author-" + authorId.LocalId;
-        return authorName;
-    }
-
-    private Task<string> ContentToText(string markup)
-        => Task.FromResult(markup); // TODO: add markup parsing
-
-    private static readonly TimeSpan BlockStartTimeGap = TimeSpan.FromSeconds(120);
-
-    private static bool IsBlockStart(ChatEntry? prevEntry, ChatEntry entry)
-    {
-        if (prevEntry == null)
-            return true;
-        if (prevEntry.AuthorId != entry.AuthorId)
-            return true;
-
-        var prevEndsAt = prevEntry.EndsAt ?? prevEntry.BeginsAt;
-        return entry.BeginsAt - prevEndsAt >= BlockStartTimeGap;
     }
 
     private async ValueTask<IReadOnlyList<ChatEntry>> LoadTailEntries(
